+++ src/IR/Function.cpp
@@ -1,5 +1,6 @@
 #include "IR/index.h"
 #include "Util/index.h"
+#include <llvm/IR/CallingConv.h>
 
 void Function::Init(Napi::Env env, Napi::Object &exports) {
     Napi::HandleScope scope(env);
@@ -25,6 +26,8 @@ void Function::Init(Napi::Env env, Napi::Object &exports) {
             InstanceMethod("setSubprogram", &Function::setSubprogram),
             InstanceMethod("getSubprogram", &Function::getSubprogram),
             InstanceMethod("getType", &Function::getType),
+            InstanceMethod("getCallingConv", &Function::getCallingConv),
+            InstanceMethod("setCallingConv", &Function::setCallingConv),
             InstanceMethod("addFnAttr", &Function::addFnAttr),
             InstanceMethod("addParamAttr", &Function::addParamAttr),
             InstanceMethod("addRetAttr", &Function::addRetAttr)
@@ -198,6 +201,19 @@ Napi::Value Function::getType(const Napi::CallbackInfo &info) {
     Napi::Env env = info.Env();
     llvm::PointerType *type = function->getType();
     return PointerType::New(env, type);
+}
+
+Napi::Value Function::getCallingConv(const Napi::CallbackInfo &info) {
+    return Napi::Number::New(info.Env(), static_cast<unsigned>(function->getCallingConv()));
+}
+
+void Function::setCallingConv(const Napi::CallbackInfo &info) {
+    Napi::Env env = info.Env();
+    if (info.Length() == 0 || !info[0].IsNumber()) {
+        throw Napi::TypeError::New(env, "setCallingConv requires a number argument");
+    }
+    auto cc = static_cast<llvm::CallingConv::ID>(info[0].As<Napi::Number>().Uint32Value());
+    function->setCallingConv(cc);
 }
 
 void Function::addFnAttr(const Napi::CallbackInfo &info) {